それぞれのViewを抽出
